+++ driver/web/conversions_user.go
@@ -188,6 +188,20 @@ func partialAccountsToDef(items []model.Account, paramsList []map[string]interfa
 	return result
 }
 
+func publicAccountToDef(item model.PublicAccount) *Def.PublicAccount {
+	identifiers := publicAccountIdentifiersToDef(item.Identifiers)
+	profile := publicProfileToDef(&item.Profile)
+	return &Def.PublicAccount{Id: item.ID, IsConnection: item.IsConnection, IsFollowing: &item.IsFollowing, Verified: &item.Verified, Profile: profile, Identifiers: identifiers}
+}
+
+func publicAccountsToDef(items []model.PublicAccount) []Def.PublicAccount {
+	result := make([]Def.PublicAccount, len(items))
+	for i, item := range items {
+		result[i] = *publicAccountToDef(item)
+	}
+	return result
+}
+
 // AccountAuthType
 func accountAuthTypeToDef(item model.AccountAuthType) Def.AccountAuthType {
 	params := item.Params
@@ -344,6 +358,18 @@ func profileFromDef(item *Def.Profile) model.Profile {
 	if item.Address != nil {
 		address = *item.Address
 	}
+	var address2 string
+	if item.Address2 != nil {
+		address2 = *item.Address2
+	}
+	var poBox string
+	if item.PoBox != nil {
+		poBox = *item.PoBox
+	}
+	var city string
+	if item.City != nil {
+		city = *item.City
+	}
 	var zipCode string
 	if item.ZipCode != nil {
 		zipCode = *item.ZipCode
@@ -367,8 +393,9 @@ func profileFromDef(item *Def.Profile) model.Profile {
 	}
 
 	return model.Profile{PhotoURL: photoURL, PronunciationURL: pronunciationURL, Pronouns: pronouns, FirstName: firstName,
-		LastName: lastName, BirthYear: int16(birthYear), Address: address, ZipCode: zipCode,
-		State: state, Country: country, Website: website, UnstructuredProperties: unstructuredProperties}
+		LastName: lastName, BirthYear: int16(birthYear), Address: address, Address2: address2,
+		POBox: poBox, City: city, ZipCode: zipCode, State: state, Country: country, Website: website,
+		UnstructuredProperties: unstructuredProperties}
 }
 
 func profileToDef(item *model.Profile) *Def.Profile {
@@ -378,9 +405,10 @@ func profileToDef(item *model.Profile) *Def.Profile {
 
 	itemVal := *item
 	birthYear := int(itemVal.BirthYear)
-	return &Def.Profile{Id: &itemVal.ID, PhotoUrl: &itemVal.PhotoURL, PronunciationUrl: &itemVal.PronunciationURL,
-		Pronouns: &itemVal.Pronouns, FirstName: &itemVal.FirstName, LastName: &itemVal.LastName, BirthYear: &birthYear,
-		Address: &itemVal.Address, ZipCode: &itemVal.ZipCode, State: &itemVal.State, Country: &itemVal.Country, Website: &itemVal.Website,
+	return &Def.Profile{Id: &itemVal.ID, PhotoUrl: &itemVal.PhotoURL, PronunciationUrl: &itemVal.PronunciationURL, Pronouns: &itemVal.Pronouns,
+		FirstName: &itemVal.FirstName, LastName: &itemVal.LastName, BirthYear: &birthYear,
+		Address: &itemVal.Address, Address2: &itemVal.Address2, PoBox: &itemVal.POBox, City: &itemVal.City, ZipCode: &itemVal.ZipCode,
+		State: &itemVal.State, Country: &itemVal.Country, Website: &itemVal.Website,
 		UnstructuredProperties: &itemVal.UnstructuredProperties}
 }
 
@@ -417,6 +445,18 @@ func profileFromDefNullable(item *Def.ProfileNullable) model.Profile {
 	if item.Address != nil {
 		address = *item.Address
 	}
+	var address2 string
+	if item.Address2 != nil {
+		address2 = *item.Address2
+	}
+	var poBox string
+	if item.PoBox != nil {
+		poBox = *item.PoBox
+	}
+	var city string
+	if item.City != nil {
+		city = *item.City
+	}
 	var zipCode string
 	if item.ZipCode != nil {
 		zipCode = *item.ZipCode
@@ -440,8 +480,36 @@ func profileFromDefNullable(item *Def.ProfileNullable) model.Profile {
 	}
 
 	return model.Profile{PhotoURL: photoURL, PronunciationURL: pronunciationURL, Pronouns: pronouns, FirstName: firstName,
-		LastName: lastName, BirthYear: int16(birthYear), Address: address, ZipCode: zipCode,
-		State: state, Country: country, Website: website, UnstructuredProperties: unstructuredProperties}
+		LastName: lastName, BirthYear: int16(birthYear), Address: address, Address2: address2,
+		POBox: poBox, City: city, ZipCode: zipCode, State: state, Country: country, Website: website,
+		UnstructuredProperties: unstructuredProperties}
+}
+
+func publicProfileToDef(item *model.PublicProfile) *Def.ProfileNullable {
+	if item == nil {
+		return nil
+	}
+
+	var birthYear int
+	if item.BirthYear != nil {
+		birthYear = int(*item.BirthYear)
+	}
+	return &Def.ProfileNullable{PhotoUrl: item.PhotoURL, PronunciationUrl: item.PronunciationURL, Pronouns: item.Pronouns,
+		FirstName: item.FirstName, LastName: item.LastName, Email: item.Email, Phone: item.Phone, BirthYear: &birthYear,
+		Address: item.Address, ZipCode: item.ZipCode, State: item.State, Country: item.Country, Website: item.Website,
+		UnstructuredProperties: &item.UnstructuredProperties}
+}
+
+func publicAccountIdentifierToDef(item model.PublicAccountIdentifier) Def.PublicAccountIdentifier {
+	return Def.PublicAccountIdentifier{Identifier: item.Identifier, Code: item.Code}
+}
+
+func publicAccountIdentifiersToDef(items []model.PublicAccountIdentifier) []Def.PublicAccountIdentifier {
+	result := make([]Def.PublicAccountIdentifier, len(items))
+	for i, item := range items {
+		result[i] = publicAccountIdentifierToDef(item)
+	}
+	return result
 }
 
 func privacyToDef(item *model.Privacy) *Def.Privacy {
@@ -450,14 +518,14 @@ func privacyToDef(item *model.Privacy) *Def.Privacy {
 	}
 
 	return &Def.Privacy{
-		Public:          &item.Public,
-		FieldVisibility: &item.FieldVisibility,
+		Public:          item.Public,
+		FieldVisibility: item.FieldVisibility,
 	}
 }
 
 func privacyFromDef(item *Def.Privacy) model.Privacy {
 	if item == nil {
-		return model.Privacy{}
+		return defaultPrivacy()
 	}
 
 	var public bool
@@ -469,12 +537,12 @@ func privacyFromDef(item *Def.Privacy) model.Privacy {
 		fieldVisibility = *item.FieldVisibility
 	}
 
-	return model.Privacy{Public: public, FieldVisibility: fieldVisibility}
+	return model.Privacy{Public: &public, FieldVisibility: &fieldVisibility}
 }
 
 func privacyFromDefNullable(item *Def.PrivacyNullable) model.Privacy {
-	if item == nil {
-		return model.Privacy{}
+	if item == nil { //default privacy
+		return defaultPrivacy()
 	}
 
 	var public bool
@@ -486,7 +554,14 @@ func privacyFromDefNullable(item *Def.PrivacyNullable) model.Privacy {
 		fieldVisibility = *item.FieldVisibility
 	}
 
-	return model.Privacy{Public: public, FieldVisibility: fieldVisibility}
+	return model.Privacy{Public: &public, FieldVisibility: &fieldVisibility}
+}
+
+func defaultPrivacy() model.Privacy {
+	return model.Privacy{
+		Public:          nil,
+		FieldVisibility: nil,
+	}
 }
 
 // MFA